chore: add page rename test
