+++ e2e-tests/history.spec.ts
@@ -1,6 +1,6 @@
 import { expect } from '@playwright/test'
 import { test } from './fixtures'
-import { createRandomPage, modKey, searchAndJumpToPage } from './utils'
+import { createRandomPage, modKey, searchAndJumpToPage, renamePage, randomString } from './utils'
 
 test('undo/redo on a page should work as expected', async ({ page, block }) => {
   const page1 = await createRandomPage(page)
@@ -64,6 +64,20 @@ test('undo/redo of a block should be allowed on page context', async ({ page, bl
 
   await searchAndJumpToPage(page, page1)
   await page.waitForTimeout(500)
+
+  await page.keyboard.press(modKey + '+z')
+  await page.waitForTimeout(100)
+
+  await expect(page.locator('text="text 1"')).toHaveCount(0)
+})
+
+test('undo/redo of a renamed page should be preserved', async ({ page, block }) => {
+  const page1 = await createRandomPage(page)
+
+  await block.mustType('text 1')
+  await page.waitForTimeout(500) // Wait for 500ms autosave period to expire
+
+  await renamePage(page, randomString(10))
 
   await page.keyboard.press(modKey + '+z')
   await page.waitForTimeout(100)
